+++ i18n-config/plugins/imaginem-builder-r2/qtx-admin.js
@@ -1,14 +1,14 @@
 (function ($) {
-    $(document).on('qtxLoadAdmin:aqpb-post', function (evt, qtx) {
+    wp.hooks.addAction('qtranx.load', 'qtranx/plugins/imaginem-builder-r2', function () {
         if (!$.fn.wp_editor)
             return;
-
+        var qtx = qTranx.hooks;
         if (!qtx.get_ml) {
             qtx.get_ml = function (h, sep) {
                 var text = h.contentField.value.trim();
-                var blocks = qtranxj_get_split_blocks(text);
-                if (!blocks || blocks.length > 1) {//already ML
-                    var contents = qtranxj_split_blocks(blocks);
+                var tokens = mlSplitRaw(text);
+                if (!tokens || tokens.length > 1) {//already ML
+                    var contents = mlParseTokens(tokens);
                     for (var lang in h.fields) {
                         h.fields[lang].value = contents[lang];
                     }