+++ i18n-config/themes/ta-pluton/qtx-admin.js
@@ -1,11 +1,9 @@
 (function ($) {
-    $(document).on('qtxLoadAdmin:ta-pluton-panel', function (evt, qtx) {
-        qtx.addLanguageSwitchAfterListener(function () {
-            $('.slide-title').each(function (i, e) {
-                var t = e.value;
-                if (!t) return;
-                $(e).parents().eq(3).find('.redux-slides-header').text(t);
-            });
+    wp.hooks.addAction('qtranx.languageSwitch', 'qtranx/themes/ta-pluton', function () {
+        $('.slide-title').each(function (i, e) {
+            var t = e.value;
+            if (!t) return;
+            $(e).parents().eq(3).find('.redux-slides-header').text(t);
         });
     });
 })(jQuery);